+++ lib/network/model/response/price_response.g.dart
@@ -7,10 +7,10 @@ part of 'price_response.dart';
 // **************************************************************************
 
 PriceResponse _$PriceResponseFromJson(Map<String, dynamic> json) =>
-    PriceResponse()
-      ..currency = json['currency'] as String?
-      ..price = _toDouble(json['price'])
-      ..xmrPrice = _toDouble(json['xmr']);
+    PriceResponse(
+      currency: json['currency'] as String?,
+      price: _toDouble(json['price']),
+    )..xmrPrice = _toDouble(json['xmr']);
 
 Map<String, dynamic> _$PriceResponseToJson(PriceResponse instance) =>
     <String, dynamic>{